RouteList: Refactor route list row
